+++ db/db_impl/db_impl.h
@@ -1082,7 +1082,7 @@ class DBImpl : public DB {
   // flush to complete, but delete the column family handle before the wait
   // finishes. For example in CompactRange.
   Status TEST_AtomicFlushMemTables(
-      const autovector<ColumnFamilyData*>& candidate_cfds,
+      const autovector<ColumnFamilyData*>& provided_candidate_cfds,
       const FlushOptions& flush_opts);
 
   // Wait for background threads to complete scheduled work.
@@ -1888,23 +1888,24 @@ class DBImpl : public DB {
   Status SwitchMemtable(ColumnFamilyData* cfd, WriteContext* context);
 
   // Select and output column families qualified for atomic flush in
-  // `selected_cfds`. If `candidate_cfds` is non-empty, it will be used as
-  // candidate CFs to select qualified ones from. Otherwise, all column families
-  // are used as candidate to select from.
+  // `selected_cfds`. If `provided_candidate_cfds` is non-empty, it will be used
+  // as candidate CFs to select qualified ones from. Otherwise, all column
+  // families are used as candidate to select from.
   void SelectColumnFamiliesForAtomicFlush(
       autovector<ColumnFamilyData*>* selected_cfds,
-      const autovector<ColumnFamilyData*> candidate_cfds = {});
+      const autovector<ColumnFamilyData*>& provided_candidate_cfds = {});
 
   // Force current memtable contents to be flushed.
   Status FlushMemTable(ColumnFamilyData* cfd, const FlushOptions& options,
                        FlushReason flush_reason,
                        bool entered_write_thread = false);
 
-  // Atomic-flush memtables from quanlified CFs in `candidate_cfds` (if
-  // provided) or in all column families
+  // Atomic-flush memtables from quanlified CFs among `provided_candidate_cfds`
+  // (if non-empty) or amomg all column families and atomically record the
+  // result to the MANIFEST.
   Status AtomicFlushMemTables(
       const FlushOptions& options, FlushReason flush_reason,
-      const autovector<ColumnFamilyData*>& candidate_cfds = {},
+      const autovector<ColumnFamilyData*>& provided_candidate_cfds = {},
       bool entered_write_thread = false);
 
   // Wait until flushing this column family won't stall writes